refactor(shim): Move node CustomEvent polyfill to Platform.shim
