+++ src/platform/polyfills/node-custom-event.ts
@@ -1,0 +1,13 @@
+// See https://github.com/nodejs/node/issues/40678#issuecomment-1126944677
+export default class CustomEvent extends Event {
+  #detail;
+
+  constructor(type: string, options?: CustomEventInit<any[]>) {
+    super(type, options);
+    this.#detail = options?.detail ?? null;
+  }
+
+  get detail() {
+    return this.#detail;
+  }
+}
